backend/requirements.txt
# ── Web Framework ──
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ── Database ──
sqlalchemy[asyncio]>=2.0.27
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# ── Celery ──
celery[redis]>=5.3.0
redis>=5.0.0

# ── Auth ──
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# ── HTTP ──
httpx>=0.27.0
aiohttp>=3.9.0

# ── ML ──
scikit-learn>=1.4.0
xgboost>=2.0.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0

# ── Graph DB ──
neo4j>=5.17.0

# ── Search ──
elasticsearch>=8.12.0

# ── NLP / Matching ──
rapidfuzz>=3.6.0
packaging>=23.0

# ── Validation ──
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# ── Utils ──
python-dateutil>=2.8.0
python-dotenv>=1.0.0
tenacity>=8.2.0
